app/pyproject.toml
[tool.poetry]
name = "django-poetry-app"
version = "0.1.0"
description = "A Django app for managing events and question aliases"
authors = ["abhiguptag22 <abhigupta.g22@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.3"
celery = "^5.4.0"
redis = "^5.2.0"
django-celery-results = "^2.5.1"
flower = "^2.0.1"
xhtml2pdf = "^0.2.16"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
